Move util to controller.
It is better place for such utility functions.
