+++ pkg/controller/util_test.go
@@ -1,4 +1,4 @@
-package connection
+package controller
 
 import (
 	"reflect"
@@ -7,6 +7,10 @@ import (
 	"github.com/container-storage-interface/spec/lib/go/csi/v0"
 	"k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+const (
+	driverName = "foo/bar"
 )
 
 func TestGetNodeID(t *testing.T) {
